+++ src/main/java/mcjty/rftools/blocks/teleporter/GuiMatterReceiver.java
@@ -146,7 +146,7 @@ public class GuiMatterReceiver extends GenericGuiContainer<MatterReceiverTileEnt
         enableButtons();
 
         drawWindow();
-        int currentRF = GenericEnergyStorageTileEntity.getCurrentRF();
+        long currentRF = GenericEnergyStorageTileEntity.getCurrentRF();
         energyBar.setValue(currentRF);
         tileEntity.requestRfFromServer(RFTools.MODID);
     }
